# app/Features/Order/Controllers/OrderReserveApiController.php
<?php

namespace App\Features\Order\Controllers;

use App\Http\Controllers\Controller;
use App\Features\Order\Services\OrderQueryService;
use Illuminate\Http\Request;
use App\Features\Order\Enums\OrderStatus;
use App\Features\BusinessHour\Services\BusinessHourService;
use Carbon\Carbon;
use App\Exceptions\BusinessException;

class OrderReserveApiController extends Controller
{
    public function __construct(
        protected OrderQueryService $orderQueryService, 
        protected BusinessHourService $businessHourService
    ) {}

    /**
     * Update reserve time for an unpaid order.
     *
     * @OA\Put(
     *     path="/api/orders/{orderId}/reserve-time",
     *     summary="Update reserve time for an unpaid order",
     *     tags={"Orders"},
     *     security={{"bearerAuth":{}}},
     *     @OA\Parameter(
     *         name="orderId",
     *         in="path",
     *         required=true,
     *         @OA\Schema(type="integer"),
     *         description="ID of the order"
     *     ),
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"reserve_time"},
     *             @OA\Property(
     *                 property="reserve_time",
     *                 type="string",
     *                 example="2025-08-05 18:30",
     *                 description="New reserve time, format: Y-m-d H:i"
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Reserve time updated successfully"
     *     ),
     *     @OA\Response(
     *         response=400,
     *         description="Validation error"
     *     ),
     *     @OA\Response(
     *         response=403,
     *         description="Forbidden - Order status is invalid for this action"
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Order not found"
     *     ),
     *    @OA\Response(
     *        response=422,
     *       description="Unprocessable Entity - The reserved time is not available"
     *    ),
     * )
     */
    public function updateReserveTime(Request $request, $orderId)
    {
        $customerId = $this->getAuthenticatedCustomer()->id;
        $order = $this->orderQueryService->getOrderById($orderId, $customerId);

        if (!$order) {
            throw new BusinessException('Order not found', 'ORDER_NOT_FOUND', 404);
        }

        if ($order->status != OrderStatus::Unpaid) {
            throw new BusinessException('Only unpaid orders can update reserve time', 'INVALID_ORDER_STATUS', 403);
        }

        $request->validate([
            'reserve_time' => 'required|date_format:Y-m-d H:i',
        ]);

        $reserveTime = Carbon::parse($request->input('reserve_time'));

        $orderType = $order->order_type;
        
        if (!$this->businessHourService->isTimeAvailableForDate($orderType, $reserveTime)) {
            throw new BusinessException('The reserved time is not available.', 'RESERVE_TIME_UNAVAILABLE', 422);
        }

        $order->reserve_time = $reserveTime->format('Y-m-d H:i');
        $order->save();

        return response()->json([
            'success' => true,
            'order_id' => $order->id,
            'reserve_time' => $order->reserve_time,
        ]);
    }
}